Add files via upload
The above code asks the user to enter their email to subscribe to the newsletter. The user's input is checked against REGEX i.e. object of the class Pattern in Java. If it matches the pattern of the email then it is a valid input, if it is not valid then the program ends with the error. This ensures that the user enters the correct email address and our database has legitimate user's email IDs.